requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
PyCryptodome
cryptography
Flask-Mail
Flask-Migrate
Werkzeug==2.3.7
